backend/celery_service/requirements.txt
celery==5.4.0
redis==5.2.1
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.6
sqlalchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
fastapi==0.115.11
jinja2==3.1.6
pytest==8.3.5
pytest-asyncio==0.25.3
flower==2.0.1
pylint==3.2.5
black==24.1.1 